+++ src/handlers/handleDeclaration/index.js
@@ -15,6 +15,7 @@ import { handleZIndex } from '../properties/handleZIndex';
 import { handleFloat } from '../properties/handleFloat';
 import { handleBorderRadiuses } from '../properties/handleBorderRadiuses';
 import { handleWidth } from '../properties/handleWidth';
+import { handleHeight } from '../properties/handleHeight';
 import { handleLetterSpacing } from '../properties/handleLetterSpacing';
 import { handlePerformanceHacks } from '../properties/handlePerformanceHacks';
 import { handleFunctions } from '../handleFunctions';
@@ -139,6 +140,10 @@ export function handleDeclaration(decl, report, options) {
 
 	if (prop === 'width' && options.widths) {
 		handleWidth(decl, report);
+	}
+
+	if (prop === 'height' && options.heights) {
+		handleHeight(decl, report);
 	}
 
 	if (prop === 'letter-spacing' && options.letterSpacings) {