+++ src/handlers/properties/handleHeight/index.js
@@ -1,0 +1,6 @@
+import { countUsage } from '../../../calculators/countUsage';
+
+export function handleHeight(decl, report) {
+	report.heights.total++;
+	countUsage(decl.value, report.heights.usage);
+}